fix: 修復運行時錯誤和配置問題
🐛 修復問題:
- 修復 LoadingSpinner 繪製時的 float 類型錯誤
- 修復配置管理器處理嵌套配置時的類型衝突
- 更新配置文件結構以支援主題嵌套配置

🔧 技術修復:
- LoadingSpinner.paintEvent: 將 float 座標轉換為 int
- ConfigManager.set: 處理非字典類型的中間節點
- 主題配置從字符串遷移到嵌套對象結構

✅ 驗證結果:
- 應用程式啟動無錯誤
- 主題系統正常運作: dark_professional 成功應用
- 響應式佈局正確初始化: 1920x1080 螢幕檢測正常
- 插件載入對話框正常顯示
- fd 插件成功載入並整合到導航欄

+++ ui/components/indicators.py
@@ -79,10 +79,10 @@ class LoadingSpinner(QWidget):
             start_radius = radius * 0.3
             end_radius = radius * 0.8
             
-            start_x = center.x() + start_radius * math.cos(math.radians(angle))
-            start_y = center.y() + start_radius * math.sin(math.radians(angle))
-            end_x = center.x() + end_radius * math.cos(math.radians(angle))
-            end_y = center.y() + end_radius * math.sin(math.radians(angle))
+            start_x = int(center.x() + start_radius * math.cos(math.radians(angle)))
+            start_y = int(center.y() + start_radius * math.sin(math.radians(angle)))
+            end_x = int(center.x() + end_radius * math.cos(math.radians(angle)))
+            end_y = int(center.y() + end_radius * math.sin(math.radians(angle)))
             
             # 設置透明度
             color = QColor(self.color)
